Rename to WpScriptAsset.php and remove unneeded object check. (PHP forces PHP objects into JS objects by default)
